# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance API",
        "description": "Personal Finance Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/report/monthly-average": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get the user's monthly average income and expenses",
                "operationId": "efa5deae92a364184c6351bc2ef09ee7",
                "responses": {
                    "200": {
                        "description": "Monthly averages returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/report/category-comparison": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Compare expenses by category",
                "operationId": "622e579f9f6bca4e87d5c9c43a424264",
                "responses": {
                    "200": {
                        "description": "Category comparison returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/report/top-expenses": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get top 10 highest expenses",
                "operationId": "0311f2876f12dd959339fcb2f3de89e5",
                "responses": {
                    "200": {
                        "description": "Top expenses returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/user/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get the authenticated user's profile",
                "operationId": "a89d41894c20f774de86b3c94ba2fd77",
                "responses": {
                    "200": {
                        "description": "User profile returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update the authenticated user's profile",
                "operationId": "f99e71876095e5de3f67b225fc947150",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jane@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/change-password": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Change the authenticated user's password",
                "operationId": "f284a538d8242cbb9f44640ba2ed76f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "oldpass"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "newpass123"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "example": "newpass123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/delete-account": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete the authenticated user's account",
                "operationId": "a4002c496d67bb12ee0563cc637fa443",
                "responses": {
                    "200": {
                        "description": "Account deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/avatar": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Upload or update the authenticated user's avatar",
                "operationId": "4db2512d9e93870b9c9055a5f295313c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete the authenticated user's avatar",
                "operationId": "5d24deeae50cb40c1057130a757944b5",
                "responses": {
                    "200": {
                        "description": "Avatar deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}